easyjson

+++ models/service.go
@@ -1,10 +1,11 @@
 package models
 
-type Status struct {
-	Forum  int64 `json:"forum"`
-	Post   int64 `json:"post"`
-	Thread int64 `json:"thread"`
-	User   int64 `json:"user"`
+//easyjson:json
+type Status struct{
+	Forum int `json:"forum"`
+	Post int `json:"post"`
+	Thread int `json:"thread"`
+	User int `json:"user"`
 }
 
 const InitScript = `ALTER SYSTEM SET checkpoint_completion_target = '0.9';
